requirements.txt
aiohttp==3.9.5
aiohttp-sse-client==0.2.1
aiolimiter==1.1.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.6.3
async-timeout==4.0.3
attrs==23.2.0
bitarray==2.9.2
bitstring==4.2.3
black==24.4.2
boto3==1.34.145
botocore==1.34.145
build==1.2.1
cachetools==4.2.2
certifi==2024.7.4
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==2.0.12
click==8.1.7
Deprecated==1.2.14
distlib==0.3.8
environs==9.5.0
exceptiongroup==1.2.2
filelock==3.15.4
Flask==2.1.2
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
identify==2.6.0
idna==3.3
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
loguru==0.7.2
MarkupSafe==2.1.1
marshmallow==3.15.0
multidict==6.0.5
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.1
pathspec==0.12.1
peewee==3.14.10
Pillow==8.4.0
pip-tools==7.4.1
platformdirs==4.2.2
pluggy==1.5.0
pre-commit==3.7.1
psycopg2-binary==2.9.3
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
pyfiglet==1.0.2
PyNaCl==1.5.0
pyproject_hooks==1.1.0
PyQRCode==1.2.1
pyTelegramBotAPI==4.21.0
pytest==8.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-telegram-bot==13.15
pytonapi==0.3.3
pytoncenter==0.0.14
pytonconnect==0.3.1
pytoniq==0.1.39
pytoniq-core==0.1.36
pytoniq-tools==0.0.6
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
telebot==0.0.5
tomli==2.0.1
ton-client-py==1.44.3.0
tonpy==0.0.0.1.2b0
tornado==6.1
treelib==1.7.0
typing_extensions==4.12.2
tzlocal==5.2
urllib3==1.26.19
virtualenv==20.26.3
websockets==12.0
Werkzeug==2.1.2
wrapt==1.16.0
x25519==0.0.2
yarl==1.9.4
